// src/ts/types/types.ts
export type coinType = {
  availableSupply: number;
  exp: string[];
  icon: string;
  id: string;
  marketCap: number;
  name: string;
  price: number;
  priceBtc: 1;
  priceChange1d: number;
  priceChange1h: number;
  priceChange1w: number;
  rank: number;
  symbol: string;
  totalSupply: number;
  twitterUrl: string;
  volume: number;
  websiteUrl: string;
};